qcom-cpufreq: Boost all online CPUs when exiting suspend
In order to improve resume performance when the CPU is waking up from
suspend (i.e., deep sleep), boost all online CPUs to their respective
maximum frequencies for 20 ms. The power cost of boosting everything to
the max for 20 ms is negligible, so this doesn't significantly increase
power consumption.

The boost is kicked off right before all the userspace processes are
thawed, as that's when the boost is needed most.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
Signed-off-by: LordShenron <priyanshupandeydelhi@gmail.com>
Signed-off-by: Ahmad Thoriq Najahi <tlogitechnjhi@gmail.com>